% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epPCA.R
\name{epPCA}
\alias{epPCA}
\title{Easy Peasy PCA Plots}
\usage{
epPCA(mtx, lab = "epPCAplot", ncp = 0, gradient = "default",
  rowColors = "dark grey", biplot = c(0, 0))
}
\arguments{
\item{mtx}{A matrix}

\item{lab}{A character label}

\item{ncp}{Number of components to examine. 0 (default) allows autodetection.}

\item{gradient}{A vector of colors as a 3 color gradient for displaying the component axis values from negative to positive}

\item{rowColors}{A vector of length 'r' of colors indicating group identity for each of 'r' rows of the matrix.}

\item{biplot}{A vector of 2 numbers indicating which components to use in the biplot (e.g. c(1,2)), default (0,0) will not generate a biplot figure.}
}
\description{
This function generates PCA plots from a matrix.
}
\examples{
mtx <- t(as.matrix(iris[,1:4]))
rowColors <- rep("white",length(iris$Species));
rowColors[which(iris$Species=="setosa")] <- "red"
rowColors[which(iris$Species=="virginica")] <- "blue"
rowColors[which(iris$Species=="versicolor")] <- "light grey"
library(RColorBrewer); gradient <- rev(colorRampPalette(rev(brewer.pal(11,'BrBG')))(100))
epPCA(mtx,"irisData",ncp=3,gradient=gradient,rowColors=rowColors,biplot=c(1,2))
}
\keyword{PCA}
